requirements.txt
streamlit
python-dotenv
requests
pandas
plotly
langchain
langchain-groq
langgraph